requirements.txt
absl-py==2.1.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.4.0
attrs==23.2.0
audioread==3.0.1
av==12.2.0
azure-cognitiveservices-speech==1.38.0
Babel==2.15.0
bangla==0.0.2
blinker==1.8.2
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
catalogue==2.0.10
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
coqui-tts==0.24.1
coqui-tts-trainer==0.1.4
ctranslate2==4.3.1
cycler==0.12.1
cymem==2.0.8
Cython==3.0.10
dateparser==1.1.8
decorator==5.1.1
distro==1.9.0
docopt==0.6.2
einops==0.8.0
elevenlabs==1.3.1
emoji==2.8.0
encodec==0.1.1
enum34==1.1.10
faster-whisper==1.0.2
filelock==3.15.4
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
g2pkk==0.1.2
grpcio==1.64.1
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.1
gruut-lang-en==2.0.1
gruut-lang-es==2.0.1
gruut-lang-fr==2.0.2
gTTS==2.5.1
h11==0.14.0
halo==0.0.31
hangul-romanize==0.1.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.7
inflect==7.3.1
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.4
joblib==1.4.2
jsonlines==1.2.0
kiwisolver==1.4.5
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
log-symbols==0.0.14
marisa-trie==1.2.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
murmurhash==1.0.10
networkx==2.8.8
nltk==3.8.1
num2words==0.5.13
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
onnxruntime==1.18.1
openai==1.35.10
openwakeword==0.6.0
packaging==24.1
pandas==1.5.3
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
preshed==3.0.9
protobuf==4.25.3
psutil==6.0.0
pvporcupine==1.9.5
PyAudio==0.2.14
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pynndescent==0.5.13
pyparsing==3.1.2
pypinyin==0.51.0
pysbd==0.3.4
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.90
pytz==2024.1
PyYAML==6.0.1
RealtimeSTT==0.2.0
RealTimeTTS==0.4.2
regex==2024.5.15
requests==2.32.3
rich==13.7.1
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.12.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.1
soundfile==0.12.1
soxr==0.3.7
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spinners==0.0.24
srsly==2.4.8
stanza==1.6.1
stream2sentence==0.2.3
SudachiDict-core==20240409
SudachiPy==0.6.8
sympy==1.13.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
tflite-runtime==2.14.0
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.1
torchaudio==2.3.1
tqdm==4.66.4
trainer==0.0.36
transformers==4.40.2
triton==2.3.1
TTS==0.22.0
typeguard==4.3.0
typer==0.12.3
typing_extensions==4.12.2
tzlocal==5.2
umap-learn==0.5.6
Unidecode==1.3.8
urllib3==2.2.2
wasabi==1.1.3
weasel==0.4.1
webrtcvad==2.0.10
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
